Add `PushAmount` trait to track the reason behind a push
We have the ability to push some funds to our peer when opening a
channel or splicing. This is generally used between LSPs and wallets
to pay fees for on-the-fly funding, but can also be used for many other
reasons.

We introduce a dedicated trait for that field instead of using a plain
amount: that allows carrying more data than just an amount, and provides
better insights of why some funds were pushed.

This is purely a refactoring at that point, without any logical changes.