Use resolved paths in `BlindedHop`
We use resolved paths in `BlindedHop`s, since we need to potentially
unwrap the first hop in the route if we're the introduction node. This
lets us refactor resolved paths to clearly distinguish the case where
we are the introduction node, for which we need to include a blinding
in our `update_add_htlc` message.