Add trick to use MPP inside blinded paths
We previously didn't support splitting a payment inside a blinded path.
This is an issue, because pairs of nodes inside blinded paths may have
multiple channels connecting them with different balance, and recipients
will include the blinded path only once in their invoice.

To allow using blinded paths multiple times for multi-part payments, we
simply duplicate them in the `BlindedRecipient`. This is a bit hacky,
but it's simple and it seems to work (it fixes the failing tests in
`OfferPaymentSpec`).