Make `Origin.Hot` a `case class`
We actually don't need to pattern match on local/channel/trampoline
for a `Hot` origin, because we instead just send back a result to
the `replyTo` actor. It's only when the origin is `Cold` that we're
interested in pattern matching on the type of payment (exclusively
done in the `PostRestartHtlcCleaner`).

The only exception is to skip local origins, but it happens in only
two places so it's probably ok to have a slightly more complex pattern
matching in that case.