Allow disabling the reputation recorder
We must allow node operators to run plain blip-0004 without our
experimental reputation recorder. It's also a stop-gap in case
there is a bug in it that can be exploited.

We also refactor to introduce more types and documentation, without
changing the reputation algorithm itself.

We also fix a few issues mentioned on the main PR comments.