Select funding inputs before sending open_channel2 and splice_init
If a changeless set of inputs is found, excess funding (if any) will be added to the proposed `fundingAmount`/`fundingContribution` before sending `open_channel2`/`splice_init` respectively.

InteractiveTxFunder sets `excess_opt` with the input value in excess of what is needed to achieve the requested target feerate.

We assume our peer requires confirmed inputs. In the future we could add a heuristic for this, but it's safer to assume they want confirmed inputs.