Accept incoming liquidity requests
If liquidity ads is active (the node operator has set some funding
rates), we accept selling liquidity to whoever wants to purchase it.
This behavior can be customized by a dedicated liquidity plugin.

This is a work-in-progress: to protect against liquidity griefing, we
must ensure that utxos are never locked when funding transactions.