Add and configure maven wrapper to use maven 3.9.9
This will make it easier to control which version of maven is used to build eclair, which in turns makes deterministic builds easier, as well as
using recent compiler options (to target newer JDKs for example).
For example, even recent versions of Github runner images use an old verion of maven and there is no easy way to upgrade.