Delay considering a channel closed when seeing an on-chain spend
Issue #2437

When external channel is spent, add it to the spentChannels list instead of immediately removing it from the graph.

Remove spent channels after 12 blocks. When a newly added channel is validated, if it spends the shared output of a recently spent channel then it is a splice.

A splice updates the graph edges to preserve balance estimate information in the graph.

Channels are immediately removed from the db so they will not added to the graph if a restart occurs before 12-blocks elapse.