Don't relay buggy extra payments
If a payer is buggy and tries to pay the same invoice multiple times,
it can lead to an edge case where the recipient accepted the first one
and purchased liquidity for it, but didn't purchase additional liquidity
and thus cannot receive the duplicate payments.

In that case, the recipient may even have released the preimage, so the
LSP (or any other node in the route) could steal the entire duplicate
payment.

To prevent that, we immediately fail such buggy duplicate payments. Note
that if the first payment was fully completed, we don't fail the extra
payments and the payer could get screwed, but they're buggy so it's their
problem...