remove superfluous signature optimization
When replaying parts that were waiting for an on-the-fly funding, we
explicitly set `commit = false` to all parts but the last one, in order
to save us from signing multiple times.

But this has side effects in the case where the last part is rejected
for some reason, because previous parts stay stuck waiting for a
`CMD_SIGN` that never arrives.