Move `channel_announcement` to `Commitment`
In preparation for announcing splices on public channels, we move the
`channel_announcement` that was previously included in `DATA_NORMAL`
to the `Commitment`'s funding status. Every confirmed commitment may
be announced to the network once we support splicing, to let the rest
of the network know that the channel wasn't closed by the spending tx
but only changed its capacity.